Split module files properly
